[WIP] create img

+++ include/cub3d.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cub3d.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: adegain <adegain@student.42.fr>            +#+  +:+       +#+        */
+/*   By: ccheyrou <ccheyrou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/08 16:20:58 by ccheyrou          #+#    #+#             */
-/*   Updated: 2023/03/16 14:19:36 by adegain          ###   ########.fr       */
+/*   Updated: 2023/03/16 19:24:07 by ccheyrou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -108,8 +108,8 @@ typedef struct s_cub3d
 	t_game 	game;
 	t_img	img;
 	char	**file;
-	int		rx;
-	int		ry;
+	float	rx;
+	float	ry;
 	int		sizex;
 	int		sizey;
 }	t_cub3d;
